requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.21.0
Pillow==8.0.1
pipenv==2020.11.15
pycparser==2.20
pygame==2.0.0
pymongo==3.11.2
six==1.15.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wtf==0.1
WTForms==2.3.3
